implement Exception catching

+++ include/static/container/queue.h
@@ -3,11 +3,12 @@
 
 #include "private/type.h"
 
+/** a container providing Queue functionality */
 struct Queue {
-  unsigned back;
-  unsigned capacity;
-  void **elements;
-  unsigned front;
+  size_t back; /**< the index of the back element + 1 */
+  size_t capacity; /**< the max number of elements held */
+  void **elements; /**< the elements */
+  size_t front; /**< the index of the front element */
 };
 
 #endif
